requirements.txt
alembic==0.9.9
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.1
Mako==1.0.7
MarkupSafe==1.0
multi-key-dict==2.0.3
pbr==4.0.2
PyMySQL==0.8.0
python-dateutil==2.7.2
python-editor==1.0.3
python-jenkins==0.4.16
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.6
urllib3==1.22
Werkzeug==0.14.1
